server/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
